apps/raspi/requirements.txt
# AdaptLight Raspi Dependencies
# Install with: pip install -r requirements.txt

# Core
anthropic>=0.18.0
openai>=1.0.0
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0

# Database
supabase>=2.0.0

gpiozero>=2.0

numpy>=1.24.0

# Optional: For Replicate STT
replicate>=0.15.0

# GPIO control
gpiozero>=2.0
lgpio>=0.2.2.0

# NeoPixel LED control (SPI-based for Pi 5)
adafruit-blinka>=8.0.0
adafruit-circuitpython-neopixel-spi>=1.0.0

# Audio input for voice
pyaudio>=0.2.14

# Text-to-speech
edge-tts>=6.1.0

# Serial communication (for Arduino COB LED)
pyserial>=3.5


